nexa-ai-complete/backend/crawler/crawler_node.py
# crawler_node.py
import aiohttp
import asyncio
import hashlib
from bs4 import BeautifulSoup
import os

# Mock classes for DistributedSet and Elasticsearch
class DistributedSet:
    def __init__(self, connection_string):
        self.connection_string = connection_string
        self.local_cache = set()

    def __contains__(self, item):
        return item in self.local_cache

    def add(self, item):
        self.local_cache.add(item)

class IntelligentCrawler:
    def __init__(self):
        self.visited = DistributedSet('redis://crawler-cache')
        self.blacklist = set(['ads', 'trackers'])
        
    async def crawl(self, url):
        # Check visited
        url_hash = hashlib.sha256(url.encode()).hexdigest()
        if url_hash in self.visited or any(bad in url for bad in self.blacklist):
            print(f"Skipping {url}")
            return None
            
        try:
            async with aiohttp.ClientSession() as session:
                async with session.get(url, timeout=10) as response:
                    if response.status != 200:
                        return None
                    html = await response.text()
                    soup = BeautifulSoup(html, 'html.parser')
                    
                    # Extract structured data
                    data = {
                        'url': url,
                        'title': soup.title.string if soup.title else '',
                        'keywords': [kw.get('content') for kw in soup.select('meta[name=keywords]')],
                        'content': self.sanitize_content(soup)
                    }
                    
                    # Store in ES (Mock)
                    await self.store_in_es(data)
                    self.visited.add(url_hash)
                    return data
        except Exception as e:
            print(f"Error crawling {url}: {e}")
            return None
                
    def sanitize_content(self, soup):
        for script in soup(['script', 'style', 'iframe']):
            script.decompose()
        return ' '.join(p.get_text() for p in soup.find_all('p'))

    async def store_in_es(self, data):
        print(f"Storing data for {data['url']}")
        # Implementation for Elasticsearch storage would go here

# Example usage
if __name__ == "__main__":
    crawler = IntelligentCrawler()
    asyncio.run(crawler.crawl("https://example.com"))
