nexa-ai-complete/backend/crawler/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
apache-airflow==2.8.0
redis==5.0.1
elasticsearch==8.11.0
